Exported functions and types

Index

IterativeLearningControl.ConstrainedILC

IterativeLearningControl.GradientILC

IterativeLearningControl.HeuristicILC

IterativeLearningControl.ILCProblem

IterativeLearningControl.ILCProblem

IterativeLearningControl.ILCSolution

IterativeLearningControl.ModelFreeILC

IterativeLearningControl.OptimizationILC

IterativeLearningControl.OptimizationILC

IterativeLearningControl.compute_input

IterativeLearningControl.compute_input

IterativeLearningControl.hankel_operator

IterativeLearningControl.hv

IterativeLearningControl.ilc

IterativeLearningControl.ilc_theorem

IterativeLearningControl.init

IterativeLearningControl.init

IterativeLearningControl.mv_hankel_operator

Docstrings

IterativeLearningControl.ILCProblem — Type

ILCProblem

Fields:

r: Reference signal

Gr: Closed-loop transfer function from reference to output

Gu: Closed-loop transfer function from plant input to output

IterativeLearningControl.ILCProblem — Method

ILCProblem(; r, Gr, Gu)
ILCProblem(; r, P, C)

Construct an ILCProblem given discrete-time transfer function models of either

The closed-loop transfer functions from reference to output and from plant input to output, or

The plant and controller transfer functions

Continuous-time transfer functions can be discretized using the function ControlSystemsBase.c2d.

r: Reference trajectory

Gr: Closed-loop transfer function from reference to output

Gu: Closed-loop transfer function from plant input to output

P: Plant model

C: Controller transfer function

IterativeLearningControl.ILCSolution — Type

ILCSolution

A structure representing the solution to an ILC problem.

Fields:

Y: Plant responses. Y[i] is the response during the ith iteration

E: Errors. E[i] is the error during the ith iteration

A: ILC inputs. A[i] is the ILC input during the ith iteration.

prob: The ILCProblem that was solved

alg: The ILCAlgorithm that was used

IterativeLearningControl.OptimizationILC — Method

OptimizationILC(; ρ = 1e-3, λ = 1e-3)

Optimization-based linear ILC algorithm from Norrlöf's thesis. This algorithm applies the ILC feedforward signal directly to the plant input.

Arguments:

ρ: Penalty on feedforward control action

λ: Step size penalty

IterativeLearningControl.compute_input — Method

compute_input(prob, alg::ILCAlgorithm, workspace, a, e)

Compute the next ILC input using the learning rule

Arguments:

alg: The ILC algorithm

workspace: A workspace created by calling init on the algorithm: workspace = init(prob, alg).

a: Previous ILC input

e: Error r - y

IterativeLearningControl.hankel_operator — Method

hankel(sys::LTISystem{<:Discrete}, N::Int)

Return a matrix operator $H$ such that $Hu^T = y^T$ where y = lsim(sys, u). $H$ is a Hankel matrix containing the Markov parameters of the system (scaled impulse response).

IterativeLearningControl.hv — Method

A little helper function that takes a matrix with dimensions (nsignals, n_timepoints) and returns a reshaped vector version that is suitable for multiplying the Hankel operator obtained by calling hankel_operator or mv_hankel_operator.

IterativeLearningControl.ilc — Method

ilc(prob, alg; iters = 5, actual=prob)

Run the ILC algorithm for iters iterations. Returns a ILCSolution structure.

To manually perform ILC iterations, see the functions

init

compute_input

To simulate the effect of plant-model mismatch, one may provide a different instance of the ILCProblem using the actual keyword argument which is used to simulate the plant response. The ILC update will be performed using the plant model from prob, while simulated data will be acquired from the plant models in the actual problem.

IterativeLearningControl.ilc_theorem — Function

ilc_theorem(alg::HeuristicILC, Gc, Gcact = nothing)

Plot the stability boundary for the ILC algorithm.

Arguments:

alg: Containing the filters $Q$ and $L$

Gc: The closed-loop system from ILC signal to output. If alg.location = :ref, this is typically given by feedback(P*C) while if alg.location = :input, this is typically given by feedback(P, C).

Gcact: If provided, this is the "actual" closed-loop system which may be constructed using a different plant model than Gc. This is useful when trying to determine if the filter choises will lead to a robust ILC algorithm. Gc may be constructed using, e.g., uncertain parameters, see https://juliacontrol.github.io/RobustAndOptimalControl.jl/dev/uncertainty/ for more details.

IterativeLearningControl.init — Method

workspace = init(prob, alg)

Initialize the ILC algorithm. This function is called internally by the funciton ilc but manual iterations require the user to initialize the workspace explicitly.

IterativeLearningControl.mv_hankel_operator — Function

mv_hankel_operator(sys::LTISystem{<:Discrete}, N::Int)

Return a matrix operator $H$ such that y == reshape(H*vec(u'), :, sys.ny)' where y = lsim(sys, u). $H$ is a block Hankel matrix containing the Markov parameters of the system (scaled impulse response).

Use of this function requires the user to manually install and load the packages using JuMP, BlockArrays.